feat: icons for returning home and editing entries
