+++ backend/database/operations.go
@@ -141,31 +141,3 @@ func GetPlayInfo(db *sql.DB, gameID string, playID string) (*models.InjuryDispla
 
 	return &instance, nil
 }
-
-func GetInjuriesByType(db *sql.DB, tag string) ([]models.InjuryDisplay, error) {
-	query := "SELECT game, play_id, type, game_position, team, jersey_number, first_name, last_name FROM injuries WHERE type = ?"
-	rows, err := db.Query(query, tag)
-	if err != nil {
-		log.Printf("Error querying database: %v", err)
-		return nil, err
-	}
-	defer rows.Close()
-
-	var results []models.InjuryDisplay
-	for rows.Next() {
-		var instance models.InjuryDisplay
-		if err := rows.Scan(&instance.Game, &instance.PlayID, &instance.Type, &instance.GamePosition, &instance.Team, &instance.JerseyNumber, &instance.FirstName, &instance.LastName); err != nil {
-			log.Printf("Error scanning row: %v", err)
-			return nil, err
-		}
-		results = append(results, instance)
-	}
-
-	if err := rows.Err(); err != nil {
-		log.Printf("Row iteration error: %v", err)
-		return nil, err
-	}
-
-	return results, nil
-
-}
