--- backend/models/gamePlayKey.go
@@ -1,6 +1,0 @@
-package models
-
-type GamePlayKey struct {
-	Game   string
-	PlayID string
-}
